document and cleanup migrator
